nova: replacement of generic Genode spinlock
Turn Genode user-level spinlock into a user-level "helpinglock". This requires
support by the kernel introduced with kernel branch r5.

The commit avoids live-locks when multiple threads with SCs on different
priority levels compete for the same user-level Genode "spinlock".

Issue #986

+++ os/include/dde_kit/spin_lock.h
@@ -17,7 +17,7 @@
 /**
  * Private spin lock type
  */
-typedef volatile int dde_kit_spin_lock;
+typedef volatile unsigned long dde_kit_spin_lock;
 
 enum { DDE_KIT_SPIN_LOCK_LOCKED, DDE_KIT_SPIN_LOCK_UNLOCKED };
